requirements.txt
flask==3.0.3
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-socketio==5.3.6
requests
gtts
pydub
mysql-connector-python==8.1.0
paho-mqtt==1.6.1
gunicorn
eventlet
werkzeug
passlib[bcrypt]
urllib3
dotenv
geopy
Flask-SQLAlchemy
pymysql
python-dotenv
# pip install -r requirements.txt